concluído issue #6 #9

+++ src/dao/DvdDao.java
@@ -4,8 +4,11 @@ import java.util.List;
 
 import javax.persistence.EntityManager;
 
+import org.hibernate.Session;
+
 import jpa.ConnectionFactory;
 import model.Dvd;
+import net.sf.ehcache.hibernate.HibernateUtil;
 
 public class DvdDao {
 
@@ -15,53 +18,23 @@ public class DvdDao {
 	/*
 	 * Create
 	 */
-	public void salva(Dvd dvd) {		
+	public String salva(Dvd dvd) {
+		String status = null;
 		
 		try {
 			em.getTransaction().begin();
 			em.persist(dvd);
 			em.getTransaction().commit();
-			
+			status = "\n\t Inserido dvd com sucesso!!!";			
 		} catch (Exception e) {
 			System.err.println("Erro CREATE dvd: " +e);;
 		} finally {
 			em.close();
 		}
 		
+		return status;
 	}
-	
-	
-	
-	/*
-	 * Update
-	 */
-	public void atualiza(Dvd dvd) {
 		
-		try {
-			em.getTransaction().begin();
-			
-			if ( dvd.getId() > 0 ){
-				em.merge(dvd);
-				em.getTransaction().commit();
-			} else {
-				// Se o id informado não existir: 
-				//em.persist(dvd);
-				System.out.println("\n\n\nID não existe");
-			}
-			
-			//em.getTransaction().commit();
-			
-		} catch (Exception e) {
-			System.err.println("Erro UPDATE dvd: " +e);
-			em.getTransaction().rollback();
-			
-		} finally {
-			em.close();
-		}
-		
-	}
-	
-	
 	
 	/*
 	 * Read by Id
@@ -76,7 +49,7 @@ public class DvdDao {
 			} catch (Exception e){
 				System.err.println("Erro READ dvd: " +e);
 			} finally {
-				em.close();
+				//em.close();
 			}
 			
 		}
@@ -84,8 +57,7 @@ public class DvdDao {
 		return d;
 	}
 	
-	
-	
+		
 	/*
 	 * Read all
 	 */
@@ -106,7 +78,34 @@ public class DvdDao {
 	}
 	
 	
+	/*
+	 * Update
+	 */
+	public void atualiza(Dvd dvd) {
+		
+		try {
+			em.getTransaction().begin();
+			
+			if ( dvd.getId() > 0 ){
+				em.merge(dvd);
+				em.getTransaction().commit();
+			} else {
+				// Se o id informado não existir: 
+				//em.persist(dvd);
+				System.out.println("\n\n\nID não existe");
+			}
+			
+		} catch (Exception e) {
+			System.err.println("Erro UPDATE dvd: " +e);
+			em.getTransaction().rollback();
+			
+		} finally {
+			em.close();
+		}		
+		
+	}
 	
+		
 	/*
 	 * Delete
 	 */
@@ -115,12 +114,12 @@ public class DvdDao {
 		
 		try {
 			dvd = em.find(Dvd.class, id);
-			if ( dvd != null ){
+			if (dvd != null && dvd.isLocacao() == false){
 				em.getTransaction().begin();
 				em.remove(dvd);
 				em.getTransaction().commit();
 			} else {
-				System.out.println("\n\n\n\t Id " +id+ " de dvd não encontrado...");
+				System.err.println("\n\n\t Falha ao remover dvd, null ou possui locação");
 			}
 			
 		} catch (Exception e) {
